Fixes broken reference in migration
